Disable RENDER_DEPTHWRITE for the particle preview.

+++ plugins/particles/RenderableParticle.h
@@ -184,7 +184,7 @@ public:
 
 			// Calculate the time-dependent angle
 			// according to docs, half the quads have negative rotation speed
-			float rotFactor = _random() % 2 == 0 ? -1 : 1;
+			int rotFactor = _random() % 2 == 0 ? -1 : 1;
 			angle += rotFactor * _stage.getRotationSpeed().integrate(timeFraction);
 
 			pushQuad(particleOrigin, _stage.getSize().evaluate(timeFraction), angle);
